# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Clinic Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /patients:
    get:
      operationId: listPatients
      tags: [patients]
      summary: List all active patients
      parameters:
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: gender
          in: query
          schema:
            type: ["string", "null"]
        - name: ageFrom
          in: query
          schema:
            type: ["integer", "null"]
        - name: ageTo
          in: query
          schema:
            type: ["integer", "null"]
        - name: maritalStatus
          in: query
          schema:
            type: ["string", "null"]
        - name: nationality
          in: query
          schema:
            type: ["string", "null"]
        - name: address
          in: query
          schema:
            type: ["string", "null"]
        - name: page
          in: query
          schema:
            type: ["integer", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Patients list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientList"
    post:
      operationId: createPatient
      tags: [patients]
      summary: Register a new patient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
...
  /patients/{id}:
    get:
      operationId: getPatient
      tags: [patients]
      summary: Get a patient by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Patient
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
    patch:
      operationId: updatePatient
      tags: [patients]
      summary: Update patient data
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientUpdate"
      responses:
        "200":
          description: Updated patient
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
    delete:
      operationId: deletePatient
      tags: [patients]
      summary: Soft delete a patient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"
  /patients/{id}/restore:
    post:
      operationId: restorePatient
      tags: [patients]
      summary: Restore a soft-deleted patient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Restored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
  /patients/{id}/permanent-delete:
    delete:
      operationId: permanentDeletePatient
      tags: [patients]
      summary: Permanently delete a patient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Permanently deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"
  /patients/{id}/visits:
    get:
      operationId: getPatientVisits
      tags: [patients]
      summary: Get visit history for a patient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Visit history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Visit"

  # ─────────────────────────────────────────────
  # APPOINTMENTS
  # ─────────────────────────────────────────────
...
  /services:
    get:
      operationId: listServices
      tags: [services]
      summary: List medical services
      parameters:
        - name: branch
          in: query
          schema:
            type: ["string", "null"]
        - name: groupId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
    post:
      operationId: createService
      tags: [services]
      summary: Create a new service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
  /services/{id}:
    patch:
      operationId: updateService
      tags: [services]
      summary: Update a service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    delete:
      operationId: deleteService
      tags: [services]
      summary: Delete a service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"
  /service-groups:
    get:
      operationId: listServiceGroups
      tags: [services]
      summary: List service groups / insurance templates
      responses:
        "200":
          description: Service groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ServiceGroup"
    post:
      operationId: createServiceGroup
      tags: [services]
      summary: Create a service group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceGroupInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceGroup"

  # ─────────────────────────────────────────────
  # INVENTORY
  # ─────────────────────────────────────────────
...
components:
  schemas:
...
    SuccessMessage:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success, message]
...
    PhoneNumber:
      type: object
      properties:
        number:
          type: string
        owner:
          type: string
      required: [number]
    Patient:
      type: object
      properties:
        id:
          type: integer
        localCode:
          type: integer
        nameAr:
          type: string
        nameEn:
          type: ["string", "null"]
        gender:
          type: string
          enum: [male, female]
        dateOfBirth:
          type: ["string", "null"]
        ageYears:
          type: ["integer", "null"]
        ageMonths:
          type: ["integer", "null"]
        ageDays:
          type: ["integer", "null"]
        phones:
          type: array
          items:
            $ref: "#/components/schemas/PhoneNumber"
        homePhone:
          type: ["string", "null"]
        maritalStatus:
          type: ["string", "null"]
        nationality:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        governorate:
          type: ["string", "null"]
        birthPlace:
          type: ["string", "null"]
        occupation:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        insuranceStatus:
          type: ["string", "null"]
        referredBy:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        isDeleted:
          type: boolean
        totalVisits:
          type: integer
        lastVisitDate:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, localCode, nameAr, gender, isDeleted, totalVisits, createdAt]
    PatientInput:
      type: object
      properties:
        localCode:
          type: ["integer", "null"]
        nameAr:
          type: string
        nameEn:
          type: ["string", "null"]
        gender:
          type: string
          enum: [male, female]
        dateOfBirth:
          type: ["string", "null"]
        phones:
          type: array
          items:
            $ref: "#/components/schemas/PhoneNumber"
        homePhone:
          type: ["string", "null"]
        maritalStatus:
          type: ["string", "null"]
        nationality:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        governorate:
          type: ["string", "null"]
        birthPlace:
          type: ["string", "null"]
        occupation:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        insuranceStatus:
          type: ["string", "null"]
        referredBy:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
      required: [nameAr, gender]
    PatientUpdate:
      type: object
      properties:
        nameAr:
          type: ["string", "null"]
        nameEn:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
        dateOfBirth:
          type: ["string", "null"]
        phones:
          type: ["array", "null"]
          items:
            $ref: "#/components/schemas/PhoneNumber"
        homePhone:
          type: ["string", "null"]
        maritalStatus:
          type: ["string", "null"]
        nationality:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        governorate:
          type: ["string", "null"]
        birthPlace:
          type: ["string", "null"]
        occupation:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        insuranceStatus:
          type: ["string", "null"]
        referredBy:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
    PatientList:
      type: object
      properties:
        patients:
          type: array
          items:
            $ref: "#/components/schemas/Patient"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [patients, total, page, limit]
    Visit:
      type: object
      properties:
        id:
          type: integer
        patientId:
          type: integer
        appointmentId:
          type: ["integer", "null"]
        date:
          type: string
        services:
          type: array
          items:
            type: string
        diagnosis:
          type: ["string", "null"]
        totalFee:
          type: number
        paidAmount:
          type: number
        remainingAmount:
          type: number
        notes:
          type: ["string", "null"]
      required: [id, patientId, date, totalFee, paidAmount, remainingAmount]

    # ── Appointment ──
...
    Service:
      type: object
      properties:
        id:
          type: integer
        groupId:
          type: ["integer", "null"]
        groupName:
          type: ["string", "null"]
        branch:
          type: string
        name:
          type: string
        isVisible:
          type: boolean
        priceType:
          type: string
          enum: [fixed, variable]
        price:
          type: number
        units:
          type: integer
        patientFee:
          type: number
        durationMinutes:
          type: ["integer", "null"]
        usesConsumables:
          type: boolean
      required: [id, branch, name, isVisible, priceType, price, units, patientFee, usesConsumables]
    ServiceInput:
      type: object
      properties:
        groupId:
          type: ["integer", "null"]
        branch:
          type: string
        name:
          type: string
        isVisible:
          type: boolean
        priceType:
          type: string
          enum: [fixed, variable]
        price:
          type: number
        units:
          type: integer
        patientFee:
          type: number
        durationMinutes:
          type: ["integer", "null"]
        usesConsumables:
          type: boolean
      required: [branch, name, isVisible, priceType, price, units, patientFee, usesConsumables]
    ServiceUpdate:
      type: object
      properties:
        name:
          type: ["string", "null"]
        isVisible:
          type: ["boolean", "null"]
        priceType:
          type: ["string", "null"]
        price:
          type: ["number", "null"]
        units:
          type: ["integer", "null"]
        patientFee:
          type: ["number", "null"]
        durationMinutes:
          type: ["integer", "null"]
        groupId:
          type: ["integer", "null"]
    ServiceGroup:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [private, insurance]
        validFrom:
          type: ["string", "null"]
        validTo:
          type: ["string", "null"]
      required: [id, name, type]
    ServiceGroupInput:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [private, insurance]
        validFrom:
          type: ["string", "null"]
        validTo:
          type: ["string", "null"]
      required: [name, type]

    # ── Inventory ──
...