# lib/api-zod/src/generated/types/phoneNumber.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Clinic Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface PhoneNumber {
  number: string;
  owner?: string;
}