# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Clinic Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { ServicePriceType } from './servicePriceType';

export interface Service {
  id: number;
  /** @nullable */
  groupId?: number | null;
  /** @nullable */
  groupName?: string | null;
  branch: string;
  name: string;
  isVisible: boolean;
  priceType: ServicePriceType;
  price: number;
  units: number;
  patientFee: number;
  /** @nullable */
  durationMinutes?: number | null;
  usesConsumables: boolean;
}